Fix cyclic imports test
