+++ starknet_py/hash/casm_class_hash.py
@@ -10,7 +10,8 @@ from starknet_py.hash.compiled_class_hash_objects import (
     BytecodeSegmentStructure,
     NestedIntList,
 )
-from starknet_py.net.client_models import CasmClass, CasmClassEntryPoint
+from starknet_py.net.client_models import CasmClassEntryPoint
+from starknet_py.net.models.compiled_casm import CasmClass
 
 CASM_CLASS_VERSION = "COMPILED_CLASS_V1"
